requirements.txt
django
djangorestframework
django-cors-headers
pymongo[snappy,gssapi,srv,tls]
dnspython
mongoengine
openai
jinja2
pandas
requests
gunicorn
pytz